[Rec-IM] Admin Reporting + Email Opt + Schedule Bug Fix

+++ Gordon360/Services/RecIM/SeriesService.cs
@@ -10,6 +10,7 @@ using System.Linq;
 using System.Threading.Tasks;
 using Gordon360.Extensions.System;
 using Gordon360.Static.Names;
+using static Microsoft.EntityFrameworkCore.DbLoggerCategory;
 
 namespace Gordon360.Services.RecIM
 {
@@ -146,14 +147,43 @@ namespace Gordon360.Services.RecIM
             return series;
         }
 
-        public SeriesScheduleViewModel GetSeriesScheduleByID(int seriesID)
+        public SeriesScheduleExtendedViewModel GetSeriesScheduleByID(int seriesID)
         {
             int scheduleID = _context.Series.Find(seriesID)?.ScheduleID ?? 0;
-            return _context.SeriesSchedule.Find(scheduleID);
+            SeriesScheduleExtendedViewModel res =  _context.SeriesSchedule.Find(scheduleID);
+            var surfaceIDs = _context.SeriesSurface.Where(ss => ss.SeriesID == seriesID).Select(s => s.SurfaceID);
+            res.SurfaceIDs = surfaceIDs;
+            return res;
         }
 
         public async Task<SeriesScheduleViewModel> PutSeriesScheduleAsync(SeriesScheduleUploadViewModel seriesSchedule)
         {
+
+            if (seriesSchedule.SeriesID is int seriesID)
+            {
+                var series = _context.Series.Find(seriesID);
+                //update surfaces
+                if (seriesSchedule.AvailableSurfaceIDs.Count() == 0) throw new UnprocessibleEntity { ExceptionMessage = "Schedule cannot have no surfaces" };
+
+                var updatedSurfaces = seriesSchedule.AvailableSurfaceIDs.ToList();
+                var seriesSurfaces = _context.SeriesSurface.Where(s => s.SeriesID == seriesID);
+                foreach (var surface in seriesSurfaces)
+                {
+                    if (!seriesSchedule.AvailableSurfaceIDs.Any(id => id == surface.ID))
+                        _context.SeriesSurface.Remove(surface);
+                    else
+                        updatedSurfaces.Remove(surface.ID);
+                }
+
+                foreach (var surfaceID in updatedSurfaces)
+                    await _context.SeriesSurface.AddAsync(
+                        new SeriesSurface
+                        {
+                            SeriesID = seriesID,
+                            SurfaceID = surfaceID
+                        });
+
+            }
             //check for exact schedule existing
             var existingSchedule = _context.SeriesSchedule.FirstOrDefault(ss =>
                 ss.StartTime.TimeOfDay == seriesSchedule.DailyStartTime.TimeOfDay &&
@@ -175,7 +205,7 @@ namespace Gordon360.Services.RecIM
                     //if the series is deleted, throw exception
                     if (series.StatusID == 0) throw new UnprocessibleEntity { ExceptionMessage = "Series has been deleted" };
 
-                    if (series != null) return existingSchedule;
+                    if (series is null) return existingSchedule;
                     series.ScheduleID = existingSchedule.ID;
                     await _context.SaveChangesAsync();
                 }
@@ -232,7 +262,13 @@ namespace Gordon360.Services.RecIM
                 foreach (var team in seriesTeams)
                 {
                     if (!update.TeamIDs.Any(id => id == team.TeamID))
+                    {
+                        // error check (if teams that being removed are already in a match, block the patch
+                        if (_context.MatchTeam.Where(mt => mt.Match.SeriesID == seriesID).Any(mt => mt.TeamID == team.TeamID))
+                            throw new UnprocessibleEntity { ExceptionMessage = $"Team {team.ID} is already in a Match in this Series and cannot be removed." };
                         _context.SeriesTeam.Remove(team);
+                    }
+                        
                     else
                         updatedSeriesTeams.Remove(team.TeamID);
                 }
